boilerplate for service account configuration page
