+++ app/configure/page-content.tsx
@@ -1,0 +1,10 @@
+import { useTitle } from 'react-use'
+
+export default function ConfigurePageContent() {
+  useTitle('Configure')
+  return (
+    <div>
+      <i>Configureee</i>
+    </div>
+  )
+}
